Added integration tests for controller, changed incorrect delete method test from service, added exceptions for endpoint “/delete/{id}”.
Added TestSecurityConfig with securityChainFilter, which does not require JWT to access endpoints. Added TestConfig with beans that should participate in testing.

+++ src/main/java/com/example/user_api/service/UserServiceImpl.java
@@ -37,7 +37,6 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User save(UserDto userDto) {
-
         return userRepository.save(userMapper.toUser(userDto));
     }
 
@@ -66,12 +65,24 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteById(Long id) {
-        userRepository.deleteById(id);
+    public ResponseEntity<String> deleteById(Long id) {
+        Optional<User> user = userRepository.findById(id);
+
+        if (user.isPresent()) {
+            userRepository.deleteById(id);
+            return new ResponseEntity<>("The user with " + id + " has been deleted",HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>("The user with id " + id + " doesn't exist", HttpStatus.NOT_FOUND);
     }
 
     @Override
     public void saveAll(List<User> users) {
         userRepository.saveAll(users);
+    }
+
+    @Override
+    public void deleteAll() {
+        userRepository.deleteAll();
     }
 }
